requirements/requirements.txt
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dash==0.39.0
dash-core-components==0.44.0
dash-html-components==0.14.0
dash-renderer==0.20.0
dash-table==3.6.0
decorator==4.3.2
et-xmlfile==1.0.1
Flask==1.0.2
Flask-Compress==1.4.0
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10
jsonschema==3.0.0
jupyter-core==4.4.0
MarkupSafe==1.1.1
nbformat==4.4.0
numpy==1.16.2
openpyxl==2.6.1
pandas==0.24.1
plotly==3.6.1
pymongo==3.7.2
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
retrying==1.3.3
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
Werkzeug==0.14.1
xlrd==1.2.0
